+++ alto/examples/ec2/artifacts.yml
