+++ alto/tests/integration/function/alto_ssm_docker.yml
@@ -1,4 +1,4 @@
-my_cloud_agent-{{ __version__ }}:
+my_cloud_agent-ssm-{{ __version__ }}:
   infra:
     type: ec2
     instance_type: t2.micro
