+++ alto/tests/integration/jupyter/alto_ssm_docker.yml
@@ -1,8 +1,8 @@
-my_cloud_agent-{{ __version__ }}:
+my_cloud_agent-ssm-{{ __version__ }}:
   infra:
     type: ec2
     instance_type: t2.micro
-    ami_image: ami-01c647eace872fc02
+    protocol: ssm
   image:
     type: docker
     base: python:3.10.8-slim-bullseye
